+++ icmake/icm-comp/parser/ppatchcontinue.c
@@ -3,7 +3,7 @@
 void p_patchContinue(SemVal *e)
 {
     e->continuelist = rss_realloc(e->continuelist,   /* expand the continuelist */
-                            (e->continuelen + 1) * sizeof(size_t));
+                            (e->continuelen + 1) * sizeof(unsigned));
 
                                             /* room for the jump-backpatch */
     e->code = rss_realloc(e->code, e->codelen += sizeof(int16_t));
