+++ icmake/icm-comp/parser/ppatchtrue.c
@@ -2,7 +2,7 @@
 
 void p_patchTrue(SemVal *e)
 {                                           /* expand the truelist */
-    e->truelist = rss_realloc(e->truelist, (e->truelen + 1) * sizeof(size_t));
+    e->truelist = rss_realloc(e->truelist, (e->truelen + 1) * sizeof(unsigned));
 
                                             /* room for the jump-backpatch */
     e->code = rss_realloc(e->code, e->codelen += sizeof(int16_t));
