+++ icmake/icm-comp/scanner/data.c
@@ -3,9 +3,9 @@
 char gs_char[2];
 char *gs_filenames;
 char *gs_savedText;
-size_t   gs_savedLineNr;
+unsigned   gs_savedLineNr;
 
-size_t gs_fileSp = N_FILES;
+unsigned gs_fileSp = N_FILES;
 FileStack gs_fileStack[N_FILES];
