+++ icmake/icm-comp/symtab/addfunction.c
@@ -8,7 +8,7 @@ int symtab_addFunction(ExprType type) /* 0: function added, -1: already def'd */
     if (symtab_findFun() != -1)
         return -1;                          /* already defined */
 
-    size_t idx = st_nextFunIdx();
+    unsigned idx = st_nextFunIdx();
 
     FunInfo *next = gs_functions.info + idx;
