+++ icmake/icm-comp/symtab/funname.c
@@ -1,6 +1,6 @@
 #include "symtab.ih"
 
-char const *symtab_funName(size_t funIdx)
+char const *symtab_funName(unsigned funIdx)
 {
     return gs_functions.info[funIdx].name;
 }
