+++ icmake/icm-comp/symtab/funnparams.c
@@ -1,6 +1,6 @@
 #include "symtab.ih"
 
-size_t symtab_fun_nParams(size_t funIdx)
+unsigned symtab_fun_nParams(unsigned funIdx)
 {
     return gs_functions.info[funIdx].nParams;
 }
