+++ icmake/icm-comp/symtab/funtype.c
@@ -1,6 +1,6 @@
 #include "symtab.ih"
 
-ExprType symtab_funType(size_t funIdx)
+ExprType symtab_funType(unsigned funIdx)
 {
     return gs_functions.info[funIdx].returnType;
 }
