+++ icmake/icm-comp/symtab/lastfunction.c
@@ -1,6 +1,6 @@
 #include "symtab.ih"
 
-size_t symtab_lastFunction() /* idx of last defined function       */
+unsigned symtab_lastFunction() /* idx of last defined function       */
 {
     return gs_functions.nFunctions - 1;
 }
