+++ icmake/icm-comp/symtab/nglobals.c
@@ -1,6 +1,6 @@
 #include "symtab.ih"
 
-size_t symtab_nGlobals()
+unsigned symtab_nGlobals()
 {
     return gs_vars.varTab->nVars;
 }
