+++ icmake/icm-comp/symtab/staddvar.c
@@ -7,7 +7,7 @@
 VarIndex st_addVar(ExprType type)
 {
     VarInfo *vi;
-    size_t idx = st_nextVarIdx(&vi);
+    unsigned idx = st_nextVarIdx(&vi);
 
     vi->name = rss_strdup(util_string());      /* set the name of the var */
     vi->type = type | e_var;          /* set the type of the var */
