+++ icmake/icm-comp/symtab/stnextvaridx.c
@@ -1,9 +1,9 @@
 #include "symtab.ih"
 
-size_t st_nextVarIdx(VarInfo **vi)  /* idx of an uninitialized record of a */
+unsigned st_nextVarIdx(VarInfo **vi)  /* idx of an uninitialized record of a */
 {                                   /* variable */
 
-    size_t idx = st_newVarIdx();    /* new variable in local or global tab */
+    unsigned idx = st_newVarIdx();    /* new variable in local or global tab */
    
                                     /* RBA a pointer to the next VarInfo */
     *vi = st_varTab(st_nestingOffset())->info + idx;
