+++ icmake/icm-comp/symtab/stparamtype.c
@@ -1,6 +1,6 @@
 #include "symtab.ih"
 
-ExprType st_paramType(size_t idx)
+ExprType st_paramType(unsigned idx)
 {
     return gs_vars.varTab[1].info[idx].type & e_typeMask;
 }
