+++ icmake/icm-comp/symtab/stvartab.c
@@ -1,6 +1,6 @@
 #include "symtab.ih"
 
-VarTab *st_varTab(size_t nesting)
+VarTab *st_varTab(unsigned nesting)
 {
     return gs_vars.varTab + (nesting > 0);
 }
