+++ icmake/icm-comp/util/scan.c
@@ -2,7 +2,7 @@
 
 void util_scan(char const *fmt, char const *text)
 {
-    size_t value;
+    unsigned value;
     sscanf(text, fmt, &value);
 
     util_printf("%u", value);
