+++ icmake/icm-exec/aux/data.c
@@ -1,4 +1,4 @@
 #include "aux.ih"
 
 char        *ga_bimname;
-size_t      ga_offset;
+unsigned      ga_offset;
